refactor(web): extract the spawn-at-a-point creation family into useNodeCreation
PR 4 of the SpatialEditor decomposition (3563 -> 3382 lines). Moves the
gesture-free creators — link, file-reference, image and group nodes,
addImageFile, and the group-the-selection verb — plus visibleCanvasRect
and panToShow into use-node-creation.ts. The seam is applyResult (the
same command sink every gesture goes through) plus a collapseExtras
callback, so creation still selects the new node exclusively through the
selection reducer. The double-press note creator stays in the editor: it
goes through the gesture reducer, not this family's idle-state results.

The four copies of the center/occupied/resolveSpawnPoint preamble
collapse into one spawnPoint helper — same expressions, one spelling.

purity-guard.test.ts's command-boundary scan also walks the new hook
(zero onChange calls today, scanned so a future direct write cannot
slip in).

Behavior-preserving: spatial-editor jsdom 43 files / 475 tests, and the
creation-covering browser suites (context-menu, group-node, image-node,
file-node-markdown, SpatialEditor.browser - 69 tests) pass un-weakened.
